From HIVE Work TestBots 5 and 6
Created Test Bots to continue GDQuest tutorial. Was able to translate through second video, but it doesn't seem to be what I'm looking for.
I tweaked things a little to better behave themselves in 4 and 5 and made sure I understood what was happening.
Rotation limit seems to exist in TestBot 3. I think it's the applied force way of doing things instead of the direct modification of the velocity.
I created TestBot 6 to start from the ground up so I can experiment with the features of CharacterBody2D and its rotations.
Current Theory, If I apply a rotation scaled down until it targets the cursor, while always moving forward toward its angle, I'll have the behavior I'm looking for and can finally move on.